FIXED: some instability issues with the max plug-ins FIXED: incorrect texts in resource file
git-svn-id: svn://svn.code.sf.net/p/nel/code/trunk@513 bf685bcc-7839-4fcd-9254-36823579bede

+++ nel/tools/3d/plugin_max/nel_patch_lib/rpo2nel.cpp
@@ -547,7 +547,7 @@ bool RPatchMesh::exportZone(INode* pNode, PatchMesh* pPM, NL3D::CZone& zone, CZo
 					// Get the tile index
 					uint tile = desc.getLayer (l).Tile;
 					uint tileRotation = desc.getLayer (l).Rotate;
-				    if (tile >= bank.getTileCount())
+				    if (tile >= (uint)bank.getTileCount())
 					{
 						std::string error = NLMISC::toString(
 							"Incorrect tileset for this zone.\r\n"
